--- _pkg.dev/wire/util/uint160_test.go
@@ -1,62 +1,0 @@
-package util
-
-import (
-	"encoding/hex"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestUInt160DecodeString(t *testing.T) {
-	hexStr := "2d3b96ae1bcc5a585e075e3b81920210dec16302"
-	val, err := Uint160DecodeString(hexStr)
-	if err != nil {
-		t.Fatal(err)
-	}
-	assert.Equal(t, hexStr, val.String())
-}
-
-func TestUint160DecodeBytes(t *testing.T) {
-	hexStr := "2d3b96ae1bcc5a585e075e3b81920210dec16302"
-	b, err := hex.DecodeString(hexStr)
-	if err != nil {
-		t.Fatal(err)
-	}
-	val, err := Uint160DecodeBytes(b)
-	if err != nil {
-		t.Fatal(err)
-	}
-	assert.Equal(t, hexStr, val.String())
-}
-
-func TestUInt160Equals(t *testing.T) {
-	a := "2d3b96ae1bcc5a585e075e3b81920210dec16302"
-	b := "4d3b96ae1bcc5a585e075e3b81920210dec16302"
-
-	ua, err := Uint160DecodeString(a)
-	if err != nil {
-		t.Fatal(err)
-	}
-	ub, err := Uint160DecodeString(b)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if ua.Equals(ub) {
-		t.Fatalf("%s and %s cannot be equal", ua, ub)
-	}
-	if !ua.Equals(ua) {
-		t.Fatalf("%s and %s must be equal", ua, ua)
-	}
-}
-
-func TestUInt160String(t *testing.T) {
-	hexStr := "b28427088a3729b2536d10122960394e8be6721f"
-	hexRevStr := "1f72e68b4e39602912106d53b229378a082784b2"
-
-	val, err := Uint160DecodeString(hexStr)
-	assert.Nil(t, err)
-
-	assert.Equal(t, hexStr, val.String())
-	assert.Equal(t, hexRevStr, val.ReverseString())
-
-}
